telegram_bot/requirements.txt
python-telegram-bot
requests
smtplib
email
csv
os
python-dotenv
